requirements.txt
numpy==1.26.4
Pillow==8.4.0
